[mars] Add services to Equipment Services + fix environment +fix EnvironmentController + fix EquipmentEnvironmentController

+++ mars/src/main/java/io/mars/basemanagement/controllers/equipment/EquipmentEnvironmentController.java
@@ -21,7 +21,7 @@ public class EquipmentEnvironmentController extends AbstractVSpringMvcController
 	@Inject
 	private EquipmentDetailController equipmentDetailController;
 
-	private final ViewContextKey<Double> TemperatureTotalMeasured = ViewContextKey.of("TemperatureTotalMeasured");
+	private final ViewContextKey<Integer> TemperatureTotalMeasured = ViewContextKey.of("TemperatureTotalMeasured");
 
 	@GetMapping("/{equipmentId}")
 	public void initContext(final ViewContext viewContext, @PathVariable("equipmentId") final Long equipmentId) {
